+++ api/client/src/main/java/org/projectnessie/client/auth/oauth2/OAuth2Client.java
@@ -15,7 +15,7 @@
  */
 package org.projectnessie.client.auth.oauth2;
 
-import static org.projectnessie.client.auth.oauth2.OAuth2ClientParams.MIN_REFRESH_DELAY;
+import static org.projectnessie.client.auth.oauth2.OAuth2ClientConfig.MIN_REFRESH_DELAY;
 import static org.projectnessie.client.auth.oauth2.TokenTypeIdentifiers.ACCESS_TOKEN;
 import static org.projectnessie.client.auth.oauth2.TokenTypeIdentifiers.REFRESH_TOKEN;
 
@@ -60,7 +60,7 @@ class OAuth2Client implements OAuth2Authenticator, Closeable {
   static final Logger LOGGER = LoggerFactory.getLogger(OAuth2Client.class);
   private static final Duration MIN_WARN_INTERVAL = Duration.ofSeconds(10);
 
-  private final OAuth2ClientParams params;
+  private final OAuth2ClientConfig config;
   private final String username;
   private final byte[] password;
   private final String scope;
@@ -75,20 +75,20 @@ class OAuth2Client implements OAuth2Authenticator, Closeable {
   private volatile Instant lastAccess;
   private volatile Instant lastWarn;
 
-  OAuth2Client(OAuth2ClientParams params) {
-    this.params = params;
-    username = params.getUsername().orElse(null);
-    password = params.getPassword().map(s -> s.getBytes(StandardCharsets.UTF_8)).orElse(null);
-    scope = params.getScope().orElse(null);
+  OAuth2Client(OAuth2ClientConfig config) {
+    this.config = config;
+    username = config.getUsername().orElse(null);
+    password = config.getAndClearPassword();
+    scope = config.getScope().orElse(null);
     tokenEndpointClient =
-        params
+        config
             .newHttpClientBuilder()
-            .setBaseUri(params.getResolvedTokenEndpoint())
-            .setAuthentication(params.getBasicAuthentication())
+            .setBaseUri(config.getResolvedTokenEndpoint())
+            .setAuthentication(config.getBasicAuthentication())
             .addResponseFilter(this::checkErrorResponse)
             .build();
-    executor = params.getExecutor();
-    lastAccess = params.getClock().get();
+    executor = config.getExecutor();
+    lastAccess = config.getClock().get();
     currentTokensStage = started.thenApplyAsync((v) -> fetchNewTokens(), executor);
     currentTokensStage
         .whenComplete((tokens, error) -> log(error))
@@ -97,7 +97,7 @@ class OAuth2Client implements OAuth2Authenticator, Closeable {
 
   @Override
   public AccessToken authenticate() {
-    Instant now = params.getClock().get();
+    Instant now = config.getClock().get();
     lastAccess = now;
     if (sleeping.compareAndSet(true, false)) {
       wakeUp(now);
@@ -132,14 +132,7 @@ class OAuth2Client implements OAuth2Authenticator, Closeable {
     return null;
   }
 
-  /**
-   * Starts the client.
-   *
-   * <p>Calling this method will trigger a first access token fetch. It will also schedule a refresh
-   * of the access token when needed.
-   *
-   * <p>Calling this method twice or more is a no-op.
-   */
+  @Override
   public void start() {
     started.complete(null);
   }
@@ -163,7 +156,7 @@ class OAuth2Client implements OAuth2Authenticator, Closeable {
             }
           }
         }
-        httpClient.close();
+        tokenEndpointClient.close();
         if (password != null) {
           Arrays.fill(password, (byte) 0);
         }
@@ -190,8 +183,8 @@ class OAuth2Client implements OAuth2Authenticator, Closeable {
   }
 
   private void maybeScheduleTokensRenewal(Tokens currentTokens) {
-    Instant now = params.getClock().get();
-    if (Duration.between(lastAccess, now).compareTo(params.getPreemptiveTokenRefreshIdleTimeout())
+    Instant now = config.getClock().get();
+    if (Duration.between(lastAccess, now).compareTo(config.getPreemptiveTokenRefreshIdleTimeout())
         > 0) {
       sleeping.set(true);
       LOGGER.debug("Sleeping...");
@@ -249,12 +242,12 @@ class OAuth2Client implements OAuth2Authenticator, Closeable {
 
   Tokens fetchNewTokens() {
     LOGGER.debug("Fetching new tokens");
-    if (params.getGrantType() == GrantType.CLIENT_CREDENTIALS) {
+    if (config.getGrantType() == GrantType.CLIENT_CREDENTIALS) {
       ClientCredentialsTokensRequest body =
           ImmutableClientCredentialsTokensRequest.builder().scope(scope).build();
       HttpResponse httpResponse = tokenEndpointClient.newRequest().postForm(body);
       return httpResponse.readEntity(ClientCredentialsTokensResponse.class);
-    } else if (params.getGrantType() == GrantType.PASSWORD) {
+    } else if (config.getGrantType() == GrantType.PASSWORD) {
       PasswordTokensRequest body =
           ImmutablePasswordTokensRequest.builder()
               .username(username)
@@ -263,12 +256,12 @@ class OAuth2Client implements OAuth2Authenticator, Closeable {
               .build();
       HttpResponse httpResponse = tokenEndpointClient.newRequest().postForm(body);
       return httpResponse.readEntity(PasswordTokensResponse.class);
-    } else if (params.getGrantType() == GrantType.AUTHORIZATION_CODE) {
-      try (AuthorizationCodeFlow flow = new AuthorizationCodeFlow(params, tokenEndpointClient)) {
+    } else if (config.getGrantType() == GrantType.AUTHORIZATION_CODE) {
+      try (AuthorizationCodeFlow flow = new AuthorizationCodeFlow(config, tokenEndpointClient)) {
         return flow.fetchNewTokens();
       }
     }
-    throw new IllegalStateException("Unsupported grant type: " + params.getGrantType());
+    throw new IllegalStateException("Unsupported grant type: " + config.getGrantType());
   }
 
   Tokens refreshTokens(Tokens currentTokens) {
@@ -294,7 +287,7 @@ class OAuth2Client implements OAuth2Authenticator, Closeable {
   }
 
   Tokens exchangeTokens(Tokens currentToken) {
-    if (!params.getTokenExchangeEnabled()) {
+    if (!config.getTokenExchangeEnabled()) {
       throw new MustFetchNewTokensException("Token exchange is disabled");
     }
     LOGGER.debug("Exchanging tokens");
@@ -310,9 +303,9 @@ class OAuth2Client implements OAuth2Authenticator, Closeable {
   }
 
   private boolean isAboutToExpire(Token token) {
-    Instant now = params.getClock().get();
-    return tokenExpirationTime(now, token, params.getDefaultRefreshTokenLifespan())
-        .isBefore(now.plus(params.getRefreshSafetyWindow()));
+    Instant now = config.getClock().get();
+    return tokenExpirationTime(now, token, config.getDefaultRefreshTokenLifespan())
+        .isBefore(now.plus(config.getRefreshSafetyWindow()));
   }
 
   /**
@@ -325,15 +318,15 @@ class OAuth2Client implements OAuth2Authenticator, Closeable {
     }
     Instant accessExpirationTime =
         tokenExpirationTime(
-            now, currentTokens.getAccessToken(), params.getDefaultAccessTokenLifespan());
+            now, currentTokens.getAccessToken(), config.getDefaultAccessTokenLifespan());
     Instant refreshExpirationTime =
         tokenExpirationTime(
-            now, currentTokens.getRefreshToken(), params.getDefaultRefreshTokenLifespan());
+            now, currentTokens.getRefreshToken(), config.getDefaultRefreshTokenLifespan());
     return shortestDelay(
         now,
         accessExpirationTime,
         refreshExpirationTime,
-        params.getRefreshSafetyWindow(),
+        config.getRefreshSafetyWindow(),
         minRefreshDelay);
   }
 
@@ -384,7 +377,7 @@ class OAuth2Client implements OAuth2Authenticator, Closeable {
         if (is != null) {
           try {
             ErrorResponse errorResponse =
-                params.getObjectMapper().readValue(is, ErrorResponse.class);
+                config.getObjectMapper().readValue(is, ErrorResponse.class);
             throw new OAuth2Exception(status, errorResponse);
           } catch (IOException ignored) {
             throw genericError(status);
@@ -399,7 +392,7 @@ class OAuth2Client implements OAuth2Authenticator, Closeable {
   }
 
   private void maybeWarn(String message, Throwable error) {
-    Instant now = params.getClock().get();
+    Instant now = config.getClock().get();
     boolean shouldWarn =
         lastWarn == null || Duration.between(lastWarn, now).compareTo(MIN_WARN_INTERVAL) > 0;
     if (shouldWarn) {